chat_with csv/requirements.txt
streamlit
cohere
langchain-community
langchain
python-dotenv
langchain_experimental
tabulate